Enumerate router instances in deploy script
The original script just did the quick thing of getting the number of
instances and counting from 0 to n-1.  This has worked fine so far, but
this PR pays the technical debt.  (BOSH doesn't guarantee that instances
will be numbered this way.)

BOSH's JSON output wasn't as useful as I hoped, so I just implemented
this with a regex.